Reorganize aro directory and switch to absolute paths
