+++ gs/frontend/aro/src/common/Navbar.tsx
@@ -1,6 +1,6 @@
-import logo from "./assets/orbital-logo.png";
+import logo from "assets/orbital-logo.png";
 import { Link } from "react-router-dom";
-import profile from "./assets/profile.png";
+import profile from "assets/profile.png";
 
 function Navbar() {
 	return (
